linkedin-scraper/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
webdriver-manager==4.0.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
lxml==4.9.3
playwright==1.40.0 